phone_service/requirements.txt
flask>=3.0.0
requests>=2.31.0
pillow>=10.0.0

